+++ include/tvm/tir/analysis.h
@@ -1,20 +1,20 @@
 /*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
+ * licensed to the apache software foundation (asf) under one
+ * or more contributor license agreements.  see the notice file
  * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
+ * regarding copyright ownership.  the asf licenses this file
+ * to you under the apache license, version 2.0 (the
+ * "license"); you may not use this file except in compliance
+ * with the license.  you may obtain a copy of the license at
  *
- *   http://www.apache.org/licenses/LICENSE-2.0
+ *   http://www.apache.org/licenses/license-2.0
  *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
+ * unless required by applicable law or agreed to in writing,
+ * software distributed under the license is distributed on an
+ * "as is" basis, without warranties or conditions of any
+ * kind, either express or implied.  see the license for the
  * specific language governing permissions and limitations
- * under the License.
+ * under the license.
  */
 
 /*!
@@ -88,60 +88,8 @@ TVM_DLL double EstimateTIRFlops(const Stmt& stmt);
 TVM_DLL double EstimateTIRFlops(const IRModule& mod);
 
 /*!
- * \brief Visitor class to perform use/def analysis, also delete unreferenced lets.
- * \sa UndefinedVars
- */
-class VarUseDefAnalyzer : public StmtExprMutator {
- public:
-  // The fields are publically readible to
-  // be accessible to the users.
-  bool visit_thread_extent_{true};
-  bool simplify_let_{true};
-  Array<Var> undefined_;
-  Array<IterVar> thread_axis_;
-  Array<PrimExpr> thread_extent_;
-  PrimExpr dyn_shmem_size_{0};
-  bool use_dyn_shmem_{false};
-  std::unordered_map<const VarNode*, int> use_count_;
-  std::unordered_map<const VarNode*, int> def_count_;
-
- private:
-  ExprDeepEqual deep_equal_;
-  std::unordered_map<Var, const LetNode*, ObjectPtrHash, ObjectPtrEqual> let_binding_;
-  Stmt VisitStmt_(const AttrStmtNode* op) final;
-
-  Stmt VisitStmt_(const LetStmtNode* op) final;
-
-  Stmt VisitStmt_(const ForNode* op) final;
-
-  Stmt VisitStmt_(const AllocateNode* op) final;
-
-  Stmt VisitStmt_(const AllocateConstNode* op) final;
-
-  Stmt VisitStmt_(const StoreNode* op) final;
-
-  Stmt VisitStmt_(const BufferStoreNode* op) final;
-
-  PrimExpr VisitExpr_(const LetNode* op) final;
-
-  PrimExpr VisitExpr_(const VarNode* op) final;
-
-  PrimExpr VisitExpr_(const ReduceNode* op) final;
-
-  PrimExpr VisitExpr_(const LoadNode* op) final;
-
-  PrimExpr VisitExpr_(const BufferLoadNode* op) final;
-
-  void HandleDef(const VarNode* v);
-
-  void HandleUse(const PrimExpr& v);
-
-  void VisitBuffer(Buffer buffer);
-};
-
-/*!
  * \brief Find undefined vars in the statement.
- * \param stmt The function to be checked.
+ * \param stmt The statement to be checked.
  * \param defs The vars that is defined.
  * \return Array of undefined vars.
  */
@@ -153,6 +101,14 @@ TVM_DLL Array<Var> UndefinedVars(const Stmt& stmt, const Array<Var>& defs);
  * \return Array of undefined vars.
  */
 TVM_DLL Array<Var> UndefinedVars(const PrimExpr& expr);
+
+/*!
+ * \brief Find undefined vars in the expression.
+ * \param stmt The statement to be checked.
+ * \param defs The vars that is defined.
+ * \return Array of undefined vars.
+ */
+TVM_DLL Array<Var> UndefinedVars(const PrimExpr& expr, const Array<Var>& defs);
 
 /*!
  * \brief Analyze the side effect
